# Conflicts:
#	api/mesh/v1alpha1/mesh_insight.pb.go
#	api/mesh/v1alpha1/mesh_insight.proto
#	app/kumactl/cmd/inspect/inspect.go
#	app/kumactl/cmd/install/testdata/install-control-plane.cni-enabled.golden.yaml
#	app/kumactl/cmd/install/testdata/install-control-plane.defaults.golden.yaml
#	app/kumactl/cmd/install/testdata/install-control-plane.global.golden.yaml
#	app/kumactl/cmd/install/testdata/install-control-plane.overrides.golden.yaml
#	app/kumactl/cmd/install/testdata/install-control-plane.remote.golden.yaml
#	app/kumactl/cmd/install/testdata/install-control-plane.with-ingress.golden.yaml
#	app/kumactl/pkg/install/k8s/control-plane/helmtemplates_vfsdata.go
#	deployments/charts/kuma/templates/cp-rbac.yaml
#	docs/cmd/kumactl/HELP.md
#	pkg/api-server/definitions/all.go
#	pkg/core/resources/manager/manager.go
#	pkg/insights/resyncer.go
#	pkg/plugins/resources/k8s/native/api/v1alpha1/mesh_insight_types.go

+++ test/e2e/externalservices_kubernetes_test.go
@@ -149,6 +149,9 @@ metadata:
 	})
 
 	AfterEach(func() {
+		if ShouldSkipCleanup() {
+			return
+		}
 		err := cluster.DeleteKuma(deployOptsFuncs...)
 		Expect(err).ToNot(HaveOccurred())
 
@@ -171,7 +174,7 @@ metadata:
 			"curl", "-v", "-m", "3", "--fail", "http://external-service.mesh")
 		Expect(err).ToNot(HaveOccurred())
 		Expect(stderr).To(ContainSubstring("HTTP/1.1 200 OK"))
-		Expect(stdout).ToNot(ContainSubstring("HTTPS"))
+		Expect(stdout).ToNot(ContainSubstring("externalservice-https-server"))
 	})
 
 	It("should route to external-service over tls", func() {
@@ -182,7 +185,7 @@ metadata:
 			"curl", "-v", "-m", "3", "--fail", "http://external-service.mesh")
 		Expect(err).ToNot(HaveOccurred())
 		Expect(stderr).To(ContainSubstring("HTTP/1.1 200 OK"))
-		Expect(stdout).To(ContainSubstring("HTTPS"))
+		Expect(stdout).To(ContainSubstring("externalservice-https-server"))
 	})
 
 	It("should disable passthrough", func() {
